Standardize the response of APIs
1. Standardize the response of APIs
2. Added bloom filter to prevent non-exist ID cache penetration problem
3. Remove InputType and OutputType interfaces
4. Updated swagger annotations

+++ src/main/java/demo/etl/service/WagerService.java
@@ -3,6 +3,8 @@ package demo.etl.service;
 import demo.etl.entity.input.Wager;
 import demo.etl.repository.input.WagerRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RBloomFilter;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
@@ -13,10 +15,13 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
+@Slf4j
 @AllArgsConstructor
 public class WagerService {
 
     private final WagerRepository wagerRepository;
+
+    private final RBloomFilter<String> wagerBloomFilter;
 
     @Cacheable(value = "wagerList", key = "#root.methodName + '_' + #example.hashCode() + '_' + #page + '_' + #size")
     public Page<Wager> page(Wager example, int page, int size){
@@ -26,22 +31,43 @@ public class WagerService {
 
     @CacheEvict(value = "wagerList", allEntries = true)
     public Wager add(Wager wager){
-        return wagerRepository.save(wager);
+        Wager result = wagerRepository.save(wager);
+        // add the wager id to bloom filter in service layer
+        wagerBloomFilter.add(wager.getId());
+        return result;
     }
 
     @CachePut(value = "wager", key = "#wager.id")
     @CacheEvict(value = "wagerList", allEntries = true)
     public Wager update(Wager wager){
+        // check if the wager id exists in bloom filter
+        if(!wagerBloomFilter.contains(wager.getId())){
+            log.info("Wager id={} not found in bloom filter", wager.getId());
+            return null;
+        }
         return wagerRepository.save(wager);
     }
 
     @CacheEvict(value = "wagerList", allEntries = true)
-    public void delete(String id){
+    public boolean delete(String id){
+        // check if the wager id exists in bloom filter
+        if(!wagerBloomFilter.contains(id)){
+            log.info("Wager id={} not found in bloom filter", id);
+            return false;
+        }
+        if(!wagerRepository.findById(id).isPresent()) {
+            return false;
+        }
         wagerRepository.deleteById(id);
+        return true;
     }
 
     @Cacheable(value = "wager", key = "#id")
     public Wager get(String id){
+        if(!wagerBloomFilter.contains(id)){
+            log.info("Wager id={} not found in bloom filter", id);
+            return null;
+        }
         return wagerRepository.findById(id).orElse(null);
     }
