+++ src/main/java/demo/etl/service/WagerSummaryService.java
@@ -3,6 +3,8 @@ package demo.etl.service;
 import demo.etl.entity.output.WagerSummary;
 import demo.etl.repository.output.WagerSummaryRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RBloomFilter;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Example;
@@ -12,10 +14,13 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
+@Slf4j
 @AllArgsConstructor
 public class WagerSummaryService {
 
     private final WagerSummaryRepository wagerSummaryRepository;
+
+    private final RBloomFilter<String> wagerSummaryBloomFilter;
 
     @Cacheable(value = "wagerSummaryList", key = "#root.methodName + '_' + #example.hashCode() + '_' + #page + '_' + #size")
     public Page<WagerSummary> page(WagerSummary example, int page, int size){
@@ -25,21 +30,43 @@ public class WagerSummaryService {
 
     @CacheEvict(value = "wagerSummaryList", allEntries = true)
     public WagerSummary add(WagerSummary wagerSummary){
-        return wagerSummaryRepository.save(wagerSummary);
+        WagerSummary result = wagerSummaryRepository.save(wagerSummary);
+        // add the wager summary id to bloom filter in service layer
+        wagerSummaryBloomFilter.add(wagerSummary.getId());
+        return result;
     }
 
     @CacheEvict(value = "wagerSummaryList", allEntries = true)
     public WagerSummary update(WagerSummary wagerSummary){
+        // check if the wager summary id exists in bloom filter
+        if(!wagerSummaryBloomFilter.contains(wagerSummary.getId())){
+            log.info("Wager summary id={} not found in bloom filter", wagerSummary.getId());
+            return null;
+        }
         return wagerSummaryRepository.save(wagerSummary);
     }
 
     @CacheEvict(value = "wagerSummaryList", allEntries = true)
-    public void delete(String id){
+    public boolean delete(String id){
+        // check if the wager summary id exists in bloom filter
+        if(!wagerSummaryBloomFilter.contains(id)){
+            log.info("Wager summary id={} not found in bloom filter", id);
+            return false;
+        }
+        if(!wagerSummaryRepository.findById(id).isPresent()) {
+            return false;
+        }
         wagerSummaryRepository.deleteById(id);
+        return true;
     }
 
     @Cacheable(value = "wagerSummary", key = "#id")
     public WagerSummary get(String id){
+        // check if the wager summary id exists in bloom filter
+        if(!wagerSummaryBloomFilter.contains(id)){
+            log.info("Wager summary id={} not found in bloom filter", id);
+            return null;
+        }
         return wagerSummaryRepository.findById(id).orElse(null);
     }
 }
